also do tests re-org
